+++ ReleaseNotes.md
@@ -1,1 +1,11 @@
-Fix an issue loading the scripts
+
+# Release Notes
+
+[1.0.5]
+
+- Fix an issue loading the scripts
+
+[1.0.6]
+
+- Add support to the BasePath parameter on Validate mode
+- Fix some bugs